corrected provisioning

+++ config/setup_db_node.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 # set -e
 
@@ -8,16 +8,49 @@ red=$(tput setaf 1)
 bold=$(tput bold)
 reset=$(tput sgr0)
 
+PASSWORD='aich1oF5'
+BASEGUID=50000
 echo "192.168.10.20 racknode-1" | tee -a /etc/hosts
 echo "192.168.10.30 racknode-2" | tee -a /etc/hosts
 
 # installation of needed packages
 echo "${bold}${green}installation of needed packages${reset}"
-yum groupinstall -y "Base"
-yum groupinstall -y "Core"
-yum install -y ntp binutils libX11 compat-libcap1 libXau cpmpat-libstdc++33 libaio libaio-devel gcc libdmx
-yum install -y glibc-devel glibc ksh make libgcc sysstat libstdc++ xorg-x11-utils xorg-x11-auth libXext libXv
-yum install -y libXtst libXi libxcb libXt libXmu libXxf86misc libXxf86dga LibXxf86vm nfs-utils
+
+yum groups mark convert
+yum groupinstall -y "Base" "Core"
+
+yum install -y <<EOF
+ntp
+binutils
+libX11
+compat-libcap1
+libXau
+cpmpat-libstdc++33
+libaio
+libaio-devel
+gcc
+libdmx
+glibc-devel
+glibc
+ksh
+make
+libgcc
+sysstat
+libstdc++
+xorg-x11-utils
+xorg-x11-auth
+libXext
+libXv
+libXtst
+libXi
+libxcb
+libXt
+libXmu
+libXxf86misc
+libXxf86dga
+LibXxf86vm
+nfs-utils
+EOF
 
 # firewall settings
 echo "${bold}${green}firewall settings${reset}"
@@ -45,28 +78,29 @@ if [[ $result == "active" ]]; then
 
     systemctl disable avahi-dnsconfd
     systemctl disable avahi-daemon
+fi
 
+if [[ -e '/etc/systemd/system/dbus-org.freedesktop.Avahi.service' ]]; then
     rm '/etc/systemd/system/dbus-org.freedesktop.Avahi.service'
+fi
+if [[ -e '/etc/systemd/system/multi-user.target.wants/avahi-daemon.service' ]]; then
     rm '/etc/systemd/system/multi-user.target.wants/avahi-daemon.service'
+fi
+if [[ -e '/etc/systemd/system/sockets.target.wants/avahi-daemon.socket' ]]; then
     rm '/etc/systemd/system/sockets.target.wants/avahi-daemon.socket'
 fi
 
 # create the needed accounts and groups
 echo "${bold}${green}create the needed accounts and groups${reset}"
-groupadd --gid 54321 oinstall
-groupadd --gid 54322 dba
-groupadd --gid 54323 asmdba
-groupadd --gid 54324 asmoper
-groupadd --gid 54325 asmadmin
-groupadd --gid 54326 oper
-# optionals but recommended groups
-groupadd --gid 54327 backupdba
-groupadd --gid 54328 dgdba
-groupadd --gid 54329 kmdba
+GUID_INDEX=0
+for grp in oinstall dba asmdb asmoper asmadmin oper backupdba dgdba kmdba; do
+    ((GUID_INDEX=$GUID_INDEX+1))
+    ((GUID=$BASEGUID+$GUID_INDEX))
+    groupadd --gid $GUID $grp
+done
 
 # create the needed users
-useradd --uid 54321 --gid oinstall --groups dba,oper,asmdba,asmoper,backupdba,dgdba,kmdba oracle
-passwd oracle
-useradd --uid 54321 --gid oinstall --groups dba,oper,asmdba,asmoper grid
-passwd grid
-
+for user in oracle grid; do
+    useradd --gid oinstall --groups dba,oper,asmdba,asmoper,backupdba,dgdba,kmdba oracle
+    echo "$user:$PASSWORD" | chpasswd 
+done
